+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Fixed
 
 - Port parameter needs to be an int
+- Context datastore request with default endpoint
 
 ## v2.3.0 - 2022-09-20